dhcp: Work around for blocked unicast requests
There may be (mis?)configured firewalls that do not allow outgoing
unicast requests on the DHCP port.  In such cases the unicast RENEW
request will fail to be sent with an EPERM error.  Instead of shutting
down the client immediately, simply ignore the error in this case.

Since the client is in the RENEWING state, it can be assumed that
broadcasts are successfully getting past the firewall.  The client's
requests will fail until T2 timer fires and the client will enter into
the REBINDING state.  At which point requests will be sent via
broadcast and hopefully succeed in renewing the lease.